Fix bug of not initializing Renderer.current_shader
- use /Zi flag to generate debugging symbols to use devenv for
debugging
- use clean.bat script for cleaning files

+++ examples/rotating_quad.c
@@ -61,8 +61,8 @@ int main(void){
 		    COLOR_GREEN,
 		    COLOR_BLUE,
 		    COLOR_WHITE);
-    deg += 100.f * ctx->delta;
-    a += ctx->delta;
+    deg += 100.f * (float)ctx->delta;
+    a += (float)ctx->delta;
 
     clock_end_draw(ctx);
   }
